SPM integration

+++ ios/trustdart/Sources/trustdart/coins/XRP.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Jay on 3/11/25.
 //
+import Foundation
 import WalletCore
 
 class XRP: Coin  {
@@ -18,13 +19,13 @@ class XRP: Coin  {
             $0.destination = txData["receiverAddress"] as! String
             $0.amount = txData["amount"] as! Int64
             if (txData["memo"] != nil) {
-                $0.destinationTag = Int64(txData["memo"] as! String)!
+                $0.destinationTag = UInt32(txData["memo"] as! String)!
             }
         }
         let input = RippleSigningInput.with {
             $0.fee = txData["fee"] as! Int64
-            $0.sequence = txData["sequence"] as! Int32
-            $0.lastLedgerSequence = txData["lastLedgerSequence"] as! Int32
+            $0.sequence = txData["sequence"] as! UInt32
+            $0.lastLedgerSequence = txData["lastLedgerSequence"] as! UInt32
             $0.account = txData["senderAddress" ] as! String
             $0.privateKey = privateKey!.data
             $0.opPayment = operation
